pyproject.toml
[project]
name = "rw-django-ninja"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types>=0.7.0",
    "asgiref>=3.8.1",
    "coreapi>=2.3.3",
    "django-cors-headers>=4.4.0",
    "django-extensions>=3.2.3",
    "django-filter>=24.3",
    "django-ninja-extra>=0.21.4",
    "django-ninja-jwt>=5.3.4",
    "django-ninja>=1.3.0",
    "django-taggit>=6.1.0",
    "email-validator>=2.2.0",
    "markdown>=3.7",
    "parameterized>=0.9.0",
    "psycopg2>=2.9.9",
    "pydantic-core>=2.23.4",
    "pydantic>=2.9.2",
    "ruff>=0.6.8",
    "sqlparse>=0.5.1",
    "typing-extensions>=4.12.2",
]

[tool.uv]
dev-dependencies = [
    "black>=24.10.0",
    "ruff>=0.6.8",
]
